requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
pydantic-ai>=0.0.22
python-dotenv>=1.0.0
streamlit>=1.41.0
psutil
